large refactoring: more DRY
