+++ src/main/scala/com/github/everpeace/k8s/throttler/controller/ClusterThrottleControllerLogic.scala
@@ -30,13 +30,15 @@ trait ClusterThrottleControllerLogic {
       ns: Namespace,
       clthrottle: v1alpha1.ClusterThrottle
     ): Boolean =
-    clthrottle.isAlreadyActiveFor(pod, ns)
+    clthrottle.isAlreadyActiveFor(pod, clthrottle.isTarget(pod, ns))
+
   def isClusterThrottleInsufficientFor(
       pod: Pod,
       ns: Namespace,
       clthrottle: v1alpha1.ClusterThrottle
     ): Boolean =
-    clthrottle.isInsufficientFor(pod, ns)
+    clthrottle.isInsufficientFor(pod, clthrottle.isTarget(pod, ns))
+
   def calcNextClusterThrottleStatuses(
       targetClusterThrottles: Set[v1alpha1.ClusterThrottle],
       podsInAllNamespaces: Set[Pod],
@@ -58,9 +60,10 @@ trait ClusterThrottleControllerLogic {
         .filter(p => p.status.exists(_.phase.exists(_ == Pod.Phase.Running)))
         .toList
       runningTotal = runningPods.==>[List[ResourceAmount]].foldLeft(zeroResourceAmount)(_ add _)
-      nextStatus   = clthrottle.spec.statusFor(runningTotal, at)
+      nextStatus = clthrottle.spec.statusFor(runningTotal, at)(
+        v1alpha1.ClusterThrottle.Status.apply)
 
-      toUpdate <- if (clthrottle.needToUpdate(nextStatus)) {
+      toUpdate <- if (clthrottle.status.needToUpdateWith(nextStatus)) {
                    List(clthrottle.key -> nextStatus)
                  } else {
                    List.empty
