+++ src/main/scala/com/github/everpeace/k8s/throttler/crd/v1alpha1/ClusterThrottle.scala
@@ -19,7 +19,6 @@ package com.github.everpeace.k8s.throttler.crd.v1alpha1
 import com.github.everpeace.k8s.throttler
 import com.github.everpeace.k8s.throttler.crd.v1alpha1
 import com.github.everpeace.k8s.throttler.crd.v1alpha1.Implicits._
-import com.github.everpeace.util.Injection._
 import skuber.ResourceSpecification.Subresources
 import skuber.apiextensions.CustomResourceDefinition
 import skuber.{
@@ -34,16 +33,17 @@ import skuber.{
 
 object ClusterThrottle {
 
+  case class Selector(selectorTerms: List[SelectorItem])
+  case class SelectorItem(
+      podSelector: LabelSelector,
+      namespaceSelector: Option[LabelSelector] = None)
+
   case class Spec(
       throttlerName: String,
       selector: Selector,
       threshold: ResourceAmount,
       temporaryThresholdOverrides: List[TemporaryThresholdOverride] = List.empty)
-
-  case class Selector(selectorTerms: List[SelectorItem])
-  case class SelectorItem(
-      podSelector: LabelSelector,
-      namespaceSelector: Option[LabelSelector] = None)
+      extends v1alpha1.Spec[Selector]
 
   case class Status(
       throttled: IsResourceAmountThrottled,
@@ -55,7 +55,10 @@ object ClusterThrottle {
 
   def apply(name: String, spec: Spec) = CustomResource[Spec, Status](spec).withName(name)
 
-  trait JsonFormats extends CommonJsonFormat {
+  trait JsonFormats
+      extends CalculatedThreshold.JsonFormat
+      with TemporaryThresholdOverride.JsonFormat
+      with ResourceAmount.JsonFormat {
     import play.api.libs.functional.syntax._
     import play.api.libs.json._
     import skuber.json.format.{maybeEmptyFormatMethods, jsPath2LabelSelFormat}
@@ -84,7 +87,7 @@ object ClusterThrottle {
       Json.format[v1alpha1.ClusterThrottle.Status]
   }
 
-  trait Syntax extends CommonSyntax {
+  trait Syntax {
     import com.github.everpeace.k8s._
 
     implicit class ThrottleSelectorItemSyntax(selectorItem: SelectorItem) {
@@ -106,52 +109,12 @@ object ClusterThrottle {
       }
     }
 
-    implicit class ClusterThrottleSpecSyntax(spec: Spec) {
-      def thresholdAt(at: skuber.Timestamp): ResourceAmount = {
-        (spec.threshold, spec.temporaryThresholdOverrides).thresholdAt(at)
-      }
-
-      def statusFor(used: ResourceAmount, at: skuber.Timestamp): Status = {
-        val calculated = thresholdAt(at)
-        v1alpha1.ClusterThrottle.Status(
-          throttled = used.isThrottledFor(calculated, isThrottledOnEqual = true),
-          used = used.filterEffectiveOn(spec.threshold),
-          calculatedThreshold = Option(
-            CalculatedThreshold(calculated,
-                                at,
-                                spec.temporaryThresholdOverrides.collectParseError()))
-        )
-      }
-    }
-
     implicit class ClusterThrottleSyntax(clthrottle: ClusterThrottle) {
       def isTarget(pod: Pod, ns: Namespace): Boolean = clthrottle.spec.selector.matches(pod, ns)
-
-      def isAlreadyActiveFor(pod: Pod, ns: Namespace): Boolean = isTarget(pod, ns) && {
-        (for {
-          st        <- clthrottle.status
-          throttled = st.throttled
-        } yield throttled.isAlreadyThrottled(pod)).getOrElse(false)
-      }
-
-      def needToUpdate(other: Status): Boolean = v1alpha1.needToUpdate(clthrottle.status, other)
-
-      def isInsufficientFor(pod: Pod, ns: Namespace): Boolean = isTarget(pod, ns) && {
-        val podResourceAmount = pod.==>[ResourceAmount]
-        val used              = clthrottle.status.map(_.used).getOrElse(zeroResourceAmount)
-        val threshold = (for {
-          st         <- clthrottle.status
-          calculated <- st.calculatedThreshold
-        } yield calculated.threshold).getOrElse(clthrottle.spec.threshold)
-        val isThrottled =
-          (podResourceAmount add used).isThrottledFor(threshold, isThrottledOnEqual = false)
-        isThrottled.resourceCounts.flatMap(_.pod).getOrElse(false) || isThrottled.resourceRequests
-          .exists(_._2)
-      }
     }
   }
 
-  trait Implicits extends JsonFormats with Syntax {
+  trait ResourceDefinitions {
     implicit val clusterThrottleResourceDefinition: ResourceDefinition[ClusterThrottle] =
       ResourceDefinition[ClusterThrottle](
         group = throttler.crd.Group,
